+++ x-pack/plugins/observability_solution/investigate_app/public/pages/details/components/investigation_header/investigation_header.tsx
@@ -5,40 +5,24 @@
  * 2.0.
  */
 
-import { EuiButtonEmpty, EuiText } from '@elastic/eui';
-import { alertOriginSchema } from '@kbn/investigation-shared';
-import { ALERT_RULE_CATEGORY } from '@kbn/rule-registry-plugin/common/technical_rule_data_field_names';
+import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import React from 'react';
-import { useKibana } from '../../../../hooks/use_kibana';
 import { useInvestigation } from '../../contexts/investigation_context';
-import { useFetchAlert } from '../../hooks/use_fetch_alert';
+import { AlertDetailsButton } from './alert_details_button';
 
 export function InvestigationHeader() {
-  const {
-    core: {
-      http: { basePath },
-    },
-  } = useKibana();
-
   const { investigation } = useInvestigation();
 
-  const alertOriginInvestigation = alertOriginSchema.safeParse(investigation?.origin);
-  const alertId = alertOriginInvestigation.success ? alertOriginInvestigation.data.id : undefined;
-  const { data: alertDetails } = useFetchAlert({ id: alertId });
+  if (!investigation) {
+    return null;
+  }
 
   return (
-    <>
-      {alertDetails && (
-        <EuiButtonEmpty
-          data-test-subj="investigationDetailsAlertLink"
-          iconType="arrowLeft"
-          size="xs"
-          href={basePath.prepend(`/app/observability/alerts/${alertId}`)}
-        >
-          <EuiText size="s">{`[Alert] ${alertDetails?.[ALERT_RULE_CATEGORY]} breached`}</EuiText>
-        </EuiButtonEmpty>
-      )}
-      {investigation && <div>{investigation.title}</div>}
-    </>
+    <EuiFlexGroup direction="column" gutterSize="m" alignItems="flexStart">
+      <EuiFlexItem>
+        <AlertDetailsButton />
+      </EuiFlexItem>
+      <EuiFlexItem>{investigation.title}</EuiFlexItem>
+    </EuiFlexGroup>
   );
 }
